pyproject.toml
[project]
name = "discord-benchmark"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py[voice]>=2.4.0",
    "psutil>=6.1.1",
]
